requirements.txt
absl-py==2.0.0
ale-py==0.8.1
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
blinker==1.7.0
cachetools==5.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.9.0
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.44.1
fsspec==2023.4.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
gym==0.26.2
gym-notices==0.0.8
gym-super-mario-bros==7.3.0
gymnasium==0.29.1
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mdurl==0.1.2
mpmath==1.3.0
nes-py==8.2.1
networkx==3.0
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==9.3.0
protobuf==4.23.4
psutil==5.9.6
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygame==2.5.2
pyglet==1.5.21
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
Shimmy==1.1.0
six==1.16.0
stable-baselines3==2.1.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
Werkzeug==3.0.1
